+++ src/appleseed/renderer/kernel/rendering/final/uniformpixelrenderer.cpp
@@ -141,7 +141,7 @@ namespace
             Tile&                       tile,
             TileStack&                  aov_tiles,
             const AABB2i&               tile_bbox,
-            const size_t                pass_hash,
+            const uint32                pass_hash,
             const Vector2i&             pi,
             const Vector2i&             pt,
             AOVAccumulatorContainer&    aov_accumulators,
